V6A emotional tagging: Hill function boost sur mycelium (Richter-Levin 2003)
- E(a) = 1 + kappa * a^n / (a^n + theta^n), kappa=1, n=3, theta=0.5
- observe() accepte arousal, count += E(a) au lieu de +1
- feed_from_transcript score arousal per-message via VADER
- 6 bornes PASS: backward compat, boost significatif, Hill switch sharp
- Bug scan fixes: decay >>->/, SQLite INTEGER->REAL, fusion arousal comment

+++ engine/core/mycelium_db.py
@@ -76,7 +76,7 @@ class MyceliumDB:
             CREATE TABLE IF NOT EXISTS edges (
                 a INTEGER NOT NULL,
                 b INTEGER NOT NULL,
-                count INTEGER NOT NULL DEFAULT 0,
+                count REAL NOT NULL DEFAULT 0,
                 first_seen INTEGER NOT NULL,
                 last_seen INTEGER NOT NULL,
                 PRIMARY KEY (a, b)
@@ -85,7 +85,7 @@ class MyceliumDB:
                 a INTEGER NOT NULL,
                 b INTEGER NOT NULL,
                 form TEXT NOT NULL,
-                strength INTEGER NOT NULL,
+                strength REAL NOT NULL,
                 fused_at INTEGER NOT NULL,
                 PRIMARY KEY (a, b)
             ) WITHOUT ROWID;
